[RENAME] 외부 configuration 파일 위치 변경
