+++ src/main/java/kr/hhplus/be/server/application/point/PointService.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.application.point;
 
 import kr.hhplus.be.server.common.exception.NotFoundUserException;
 import kr.hhplus.be.server.domain.point.Point;
-import kr.hhplus.be.server.infrastructure.point.PointRepository;
+import kr.hhplus.be.server.infrastructure.persistence.point.PointRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
