+++ src/main/java/kr/hhplus/be/server/application/product/ProductService.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.application.product;
 
 import kr.hhplus.be.server.domain.product.Product;
-import kr.hhplus.be.server.infrastructure.product.ProductRepository;
+import kr.hhplus.be.server.infrastructure.persistence.product.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
