+++ src/main/java/kr/hhplus/be/server/infrastructure/persistence/point/PointRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.point;
+package kr.hhplus.be.server.infrastructure.persistence.point;
 
 import kr.hhplus.be.server.domain.point.Point;
 import org.springframework.data.jpa.repository.JpaRepository;
