+++ src/main/java/kr/hhplus/be/server/infrastructure/persistence/product/ProductRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.product;
+package kr.hhplus.be.server.infrastructure.persistence.product;
 
 import kr.hhplus.be.server.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
