+++ src/test/java/kr/hhplus/be/server/application/product/ProductServiceTest.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.application.product;
 
 import kr.hhplus.be.server.domain.product.Product;
-import kr.hhplus.be.server.infrastructure.product.ProductRepository;
+import kr.hhplus.be.server.infrastructure.persistence.product.ProductRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import static org.mockito.BDDMockito.*;
